refactor: Extract modal content
