+++ src/pages/listPage/Donation/components/DonationModalContent.module.css
@@ -1,0 +1,85 @@
+.contentContainer {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 10px;
+}
+
+.contentWrapper {
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+}
+
+.idolImg {
+  width: 158px;
+  height: 206px;
+  object-fit: cover;
+  border-radius: 8px;
+  box-shadow: 0px 0px 40px 0px rgba(255, 255, 255, 0.1);
+}
+
+.titleWrapper {
+  display: flex;
+  flex-direction: column;
+  gap: 6px;
+}
+
+.subtitle {
+  color: #fff;
+  opacity: 0.4;
+  font-size: 12px;
+  font-weight: 400;
+}
+
+.title {
+  color: var(--white-100);
+  font-size: 14px;
+  font-weight: 500;
+}
+
+.creditContainer {
+  width: 100%;
+  margin-top: 24px;
+  display: flex;
+  flex-direction: column;
+  gap: 24px;
+}
+
+.creditInputWrapper {
+  border-radius: 8px;
+  border: 1px solid #fff;
+  background: #272f3d;
+  padding: 16px;
+  display: flex;
+  gap: 8px;
+}
+
+.error {
+  border: 1px solid #ff3b3b;
+}
+
+.errorMessage {
+  color: #ff2626;
+  font-size: 12px;
+  font-weight: 500;
+  display: block;
+  margin-top: 6px;
+}
+
+.creditInput {
+  width: 100%;
+  color: #fff;
+  font-size: 20px;
+  font-weight: 700;
+  background: transparent;
+}
+
+.creditInput::placeholder {
+  color: var(--gray-300);
+}
+
+.buttonWrapper {
+  height: 40px;
+}
