+++ src/pages/listPage/Donation/components/Modal.jsx
@@ -1,19 +1,13 @@
-import Button from './Button';
 import styles from './Modal.module.css';
 import ReactDOM from 'react-dom';
 import closeIcon from '@/assets/icons/close-modal.svg';
-import creditIcon from '@/assets/icons/credit.svg';
 import { useState, useEffect } from 'react';
-import { useCredit } from '@/hooks/useCredit';
-import { toast } from 'react-toastify';
-import { proceedDonation } from '@/apis/donationsApi';
-import { donationsErrorMsg } from '@/constants/errorMessages';
-import { donationsMsg } from '@/constants/successMessages';
+
+import DonationModalContent from './DonationModalContent';
 
 export default function Modal({ isOpen, onClose, id, title, subtitle, idol }) {
   const [toDonateCredit, setToDonateCredit] = useState(0);
   const [errorMsg, setErrorMsg] = useState('');
-  const { credit, deductCredit } = useCredit();
 
   useEffect(() => {
     const escKeyModalClose = (e) => {
@@ -43,38 +37,6 @@ export default function Modal({ isOpen, onClose, id, title, subtitle, idol }) {
     onClose();
   };
 
-  const handleCreditOnChange = (e) => {
-    if (isNaN(Number(e.target.value))) {
-      e.target.value = '';
-      setErrorMsg(donationsErrorMsg.onlyNumber);
-      return;
-    }
-
-    setErrorMsg(null);
-    setToDonateCredit(e.target.value);
-  };
-
-  const handleOnSubmit = async (e) => {
-    e.preventDefault();
-
-    if (credit < toDonateCredit) {
-      setErrorMsg(donationsErrorMsg.creditNotEnough);
-      return;
-    }
-
-    try {
-      await proceedDonation(id, toDonateCredit);
-      deductCredit(toDonateCredit);
-      setErrorMsg(null);
-      setToDonateCredit(0);
-      toast.success(donationsMsg.apiCallSuccess);
-      onClose();
-    } catch (error) {
-      console.log(error);
-      toast.error(donationsErrorMsg.apiCallFailure);
-    }
-  };
-
   return ReactDOM.createPortal(
     <div className={styles.modalBackground} onClick={handleOnCloseModal}>
       <div className={styles.modalContent}>
@@ -85,44 +47,17 @@ export default function Modal({ isOpen, onClose, id, title, subtitle, idol }) {
           </button>
         </div>
         <div>
-          <section className={styles.contentContainer}>
-            <div className={styles.contentWrapper}>
-              <img
-                src={idol.profilePicture}
-                alt={idol.name}
-                className={styles.idolImg}
-              />
-              <div className={styles.titleWrapper}>
-                <span className={styles.subtitle}>{subtitle}</span>
-                <h4 className={styles.title}>{title}</h4>
-              </div>
-            </div>
-          </section>
-          <form className={styles.creditContainer} onSubmit={handleOnSubmit}>
-            <div>
-              <div
-                className={`${styles.creditInputWrapper} ${errorMsg ? styles.error : ''}`}
-              >
-                <input
-                  className={styles.creditInput}
-                  type="text"
-                  placeholder="크레딧 입력"
-                  onChange={handleCreditOnChange}
-                />
-                <img src={creditIcon} alt="크레딧 아이콘" />
-              </div>
-              {errorMsg && (
-                <span className={styles.errorMessage}>{errorMsg}</span>
-              )}
-            </div>
-            <div className={styles.buttonWrapper}>
-              <Button
-                text="후원하기"
-                type="submit"
-                disabled={!toDonateCredit}
-              />
-            </div>
-          </form>
+          <DonationModalContent
+            id={id}
+            title={title}
+            subtitle={subtitle}
+            idol={idol}
+            errorMsg={errorMsg}
+            setErrorMsg={setErrorMsg}
+            toDonateCredit={toDonateCredit}
+            setToDonateCredit={setToDonateCredit}
+            onClose={onClose}
+          />
         </div>
       </div>
     </div>,
